pyscript/requirements.txt
openai>=2.20.0
psycopg2-binary>=2.9.11
notion-client>=2.2.1
